requirements.txt
asgiref==3.7.2
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
django-leaflet==0.29.0
djangorestframework==3.14.0
Markdown==3.4.4
numpy==1.25.2
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
XlsxWriter==3.1.2
